Moved TableIdentifierComparator next to location key
